requirements.txt
torch>=2.0.0
torchvision>=0.15.0
gradio>=4.0.0
pillow>=9.0.0
numpy>=1.21.0
opencv-python-headless>=4.5.0
timm>=0.9.0
transformers>=4.20.0
matplotlib>=3.5.0
huggingface-hub>=0.15.0 